refactor(mir): extract edge-owner replay
